+++ lib/live_select/component.ex
@@ -15,6 +15,7 @@ defmodule LiveSelect.Component do
     allow_clear: false,
     available_option_class: nil,
     clear_button_class: nil,
+    clear_button_extra_class: nil,
     user_defined_options: false,
     container_class: nil,
     container_extra_class: nil,
@@ -45,7 +46,8 @@ defmodule LiveSelect.Component do
     tailwind: [
       active_option: ~W(text-white bg-gray-600),
       available_option: ~W(cursor-pointer hover:bg-gray-400 rounded),
-      clear_button: ~W(hidden),
+      clear_button: ~W(hidden cursor-pointer),
+      clear_tag_button: ~W(cursor-pointer),
       container: ~W(h-full text-black relative),
       dropdown: ~W(absolute rounded-md shadow z-50 bg-gray-100 inset-x-0 top-full),
       option: ~W(rounded px-4 py-1),
@@ -59,7 +61,8 @@ defmodule LiveSelect.Component do
     daisyui: [
       active_option: ~W(active),
       available_option: ~W(cursor-pointer),
-      clear_button: ~W(hidden),
+      clear_button: ~W(hidden cursor-pointer),
+      clear_tag_button: ~W(cursor-pointer),
       container: ~W(dropdown dropdown-open),
       dropdown:
         ~W(dropdown-content z-[1] menu menu-compact shadow rounded-box bg-base-200 p-1 w-full),
@@ -607,8 +610,7 @@ defmodule LiveSelect.Component do
     String.split(class_override)
   end
 
-  defp class(:none, element, nil, _class_extend)
-       when element not in [:clear_button, :clear_tag_button] do
+  defp class(:none, element, nil, _class_extend) do
     raise """
     When using `style: :none`, please use only `#{element}_class` and not `#{element}_extra_class`
     """
@@ -681,7 +683,7 @@ defmodule LiveSelect.Component do
       xmlns="http://www.w3.org/2000/svg"
       viewBox="0 0 20 20"
       fill="currentColor"
-      class={["w-5 h-5", @class]}
+      class={["w-5 h-5"]}
     >
       <path d="M6.28 5.22a.75.75 0 00-1.06 1.06L8.94 10l-3.72 3.72a.75.75 0 101.06 1.06L10 11.06l3.72 3.72a.75.75 0 101.06-1.06L11.06 10l3.72-3.72a.75.75 0 00-1.06-1.06L10 8.94 6.28 5.22z" />
     </svg>