+++ lib/support/live_select_web/live/showcase_live.ex
@@ -15,6 +15,7 @@ defmodule LiveSelectWeb.ShowcaseLive do
       :active_option_class,
       :available_option_class,
       :clear_button_class,
+      :clear_button_extra_class,
       :container_class,
       :container_extra_class,
       :dropdown_class,